The Y-axis position handler was not working correctly
Determined that the logic to determine the center point of the sprite was working only on the x-axis. This obviously will not do. So I have updated the handler to always work with a Vector2 so that I have access to the X and Y axes (not a typo, I had to look up the correct spelling of the plural for axis) at all times.