The background is now parallaxed!
Added a distance and speed adjuster to the parallax handler. The distance allows the sensation of parallaxing to be created, the distance is set from within the editor. There is a default of 1 in case the distance is forgotten, otherwise we would have divide by 0 issues. The speed is not actually used yet, but it is all wired up. I am thinking that in the future I could add a thrust option that increases the speed at some point, or perhaps for some in-game videos.