Finally satisfied with default firing speed
The idea is that there will be upgrades that you can purchase for the ship, and some of those upgrades will increase your fire rate. In addition I locked the frame rate to 60 frames per second in order to stabilize the rate of fire for the UI. When the frame rate is not locked you can see that the update method is not being called at the exact same time for any given amount of time.