Switched to the new input system.
Finished getting the input system to work correctly. Changed the mouse from using delta to only using its position. Had to write some handler logic to account for switching from a gamepad to a mouse and keyboard. Also I disabled the ability to use the arrow or WASD keys for movement, this was done to just simplify the logic. Also scaled the movement for the gamepad, it was far too fast.

Still ToDo: using the joystick from a control pad will result in stable motion once OnMove is triggered, but it will stop once that change is processed. Need to look into setting up a co-routine that will start the movement and wait for another call for OnMove to stop the current movement. To reproduce this issue: use a gamepad and hold the left stick in any direction. Do not allow the joystick to move past this initial point, not even a slight twitch. Notice on the screen that the ship will stop moving at some point. Then notice that movement will begin again if you change the position of the left stick.